requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Limiter==3.5.0
Werkzeug==2.3.7 
requests
openai==1.12.0
flask-mail==0.9.1
secrets